Major Update
Still no ready to use version. Added a script for automating Steam-Award-Nomination. It can get stuck, if there is an account which already voted and/or wrote a review for an owned game. Only way to circumvent this is to update the game the script tries to update the review for (Payday 2) or use another AppID on logon. One should also make sure, ALL bot-accounts own this one specific game. Nonetheless one should change all parts of the config at the top of the script.