Fix licenseBulkActivation getting stuck
Due to our custom logout being to fast, it can fail on reloading the page while changing to login and instead just refreshing the store-page for given appid. Fix should send you back to login where it gets executed further if there accounts left which need the license for the given appid.