Add support for getting cards from minigame
This is just a quick and dirty solution and you have to additionally install the userscript from coryshaw (many thanks!)[https://github.com/coryshaw1/saliens-hack] and make it active before you run the discovery-queue function on login.

This will reload the and play the games for times and potentially gets you 3 cards on every acc. I don't know if it will work for other days. It's free for mainline btw, topkek.