Script for sending massive trades
Can only be executed manually currently, but there is no good script or extension yet, which can achieve this heavy task easily. This script is able to get all assetids of steam-sale cards (non-foil), extract only full sets, returns a string with assetids which you then can use on background-page of this extension to send the trade.